+++ src/MenuItem.js
@@ -31,7 +31,10 @@ class MenuItem extends React.Component {
     let headerClass = bootstrapUtils.prefix(this.props, 'header');
 
     if (this.props.divider) {
-      return <li role="separator" className="divider" />;
+      return (
+        <li role="separator"
+          className={classnames('divider', this.props.className)} />
+      );
     }
 
     if (this.props.header) {
@@ -64,8 +67,21 @@ class MenuItem extends React.Component {
 }
 
 MenuItem.propTypes = {
+
+  /**
+   * Highlight the menu item as active.
+   */
   active: React.PropTypes.bool,
+
+  /**
+   * Disable the menu item, making it unselectable.
+   */
   disabled: React.PropTypes.bool,
+
+  /**
+   * Styles the menu item as a horizontal rule, providing visual separation between
+   * groups of menu items.
+   */
   divider: all(
     React.PropTypes.bool,
     props => {
@@ -74,14 +90,51 @@ MenuItem.propTypes = {
       }
     }
   ),
+
+  /**
+   * Value passed to the `onSelect` handler, useful for identifying the selected menu item.
+   */
   eventKey: React.PropTypes.any,
+
+  /**
+   * Styles the menu item as a header label, useful for describing a group of menu items.
+   */
   header: React.PropTypes.bool,
+
+  /**
+   * HTML `href` attribute corresponding to `a.href`.
+   */
   href: React.PropTypes.string,
+
+  /**
+   * HTML `target` attribute corresponding to `a.target`.
+   */
   target: React.PropTypes.string,
+
+  /**
+   * HTML `title` attribute corresponding to `a.title`.
+   */
   title: React.PropTypes.string,
+
+  /**
+   * Callback fired when the menu item is clicked.
+   */
   onClick: React.PropTypes.func,
+
   onKeyDown: React.PropTypes.func,
+
+  /**
+   * Callback fired when the menu item is selected.
+   *
+   * ```js
+   * function(Object event, Any eventKey)
+   * ```
+   */
   onSelect: React.PropTypes.func,
+
+  /**
+   * HTML `id` attribute.
+   */
   id: React.PropTypes.oneOfType([
     React.PropTypes.string,
     React.PropTypes.number