+++ test/NavbarSpec.js
@@ -256,4 +256,17 @@ describe('Navbar', () => {
     ReactTestUtils.findRenderedDOMComponentWithClass(instance, 'my-navbar-nav');
     ReactTestUtils.findRenderedDOMComponentWithClass(instance, 'my-navbar-right');
   });
+
+  it('Should add custom className to header', () => {
+    let instance = ReactTestUtils.renderIntoDocument(
+      <Navbar>
+        <Navbar.Header className='test'>
+          <Navbar.Brand />
+        </Navbar.Header>
+      </Navbar>
+    );
+
+    ReactTestUtils.findRenderedDOMComponentWithClass(instance, 'test');
+  });
+
 });